crq_project/requirements.txt
django
pandas
pyodbc
pybase64
openpyxl
python-dotenv
django-cors-headers
django-session-timeout
openai==0.28
python-docx